refactor(api): give Firestore helpers a submodule
`repositories/` held per-domain subdirectories next to flat files
carrying Firestore-wide concerns, so the shared infrastructure had no
home of its own and paid for one with a `firestore-` filename prefix.
`repositories/firestore/` is that home, and the prefix goes with it.

`snapshot.ts` moves alongside `error.ts`: it takes a Firestore
`DocumentSnapshot` and is exactly the "converter" the directory exists
to hold. `schema-version.ts` stays flat — it versions our stored
documents through verzod and imports nothing from Firestore.

`firestoreErrorToHttpException` keeps its prefix. The path carries the
context inside the module, but the only consumer is `app.onError`,
which discriminates among error origins; a bare `errorToHttpException`
would read there as the one error mapper rather than the Firestore one.
