+++ apps/api/src/repositories/firestore/error.ts
